Added booking and opening models, services, entities and controllers.

+++ LandonApi/Infrastructure/MappingProfile.cs
@@ -8,7 +8,7 @@ namespace LandonApi.Infrastructure
         public MappingProfile()
         {
             CreateMap<RoomEntity, Room>()
-                .ForMember(dest => dest.Rate, opt => opt.MapFrom(src => src.Rate / 100.0))
+                .ForMember(dest => dest.Rate, opt => opt.MapFrom(src => (decimal)(src.Rate / 100.0)))
                 .ForMember(dest => dest.Self, opt => opt.MapFrom(src =>
                 Link.To(
                     nameof(Controllers.RoomsController.GetRoomById),
